+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## [0.3.0] - ????-??-??
 ### Added
- - `Request` class
+ - `ServerRequest` class
 ### Removed
- - Request helpers - use the new `Request` wrapper class instead
+ - Request helpers - use the new `ServerRequest` wrapper class instead
 
 ## [0.2.0] - 2018-11-16
 ### Added